Updated the clean and commit scripts

+++ clean.sh
@@ -2,9 +2,11 @@
 
 rm cmake_install.cmake
 rm CMakeCache.txt
-cd ./CMakeFiles
-rm -r ./*
-cd ../
-rmdir ./CMakeFiles
+if [ -d "CMakeFiles/" ]; then
+	cd ./CMakeFiles
+	rm -r ./*
+	cd ..
+	rmdir ./CMakeFiles
+fi
 cd ./build
 rm -r ./*
